+++ src/rule-definitions.ts
@@ -2,6 +2,7 @@ import type { RuleDefinition } from '#src/common/rule.js';
 import fieldNameMappingSnakeCase from '#src/rules/field-name-mapping-snake-case.js';
 import fieldOrder from '#src/rules/field-order.js';
 import forbidField from '#src/rules/forbid-field.js';
+import forbidRequiredIgnoredField from '#src/rules/forbid-required-ignored-field.js';
 import modelNameGrammaticalNumber from '#src/rules/model-name-grammatical-number.js';
 import modelNameMappingSnakeCase from '#src/rules/model-name-mapping-snake-case.js';
 import modelNamePrefix from '#src/rules/model-name-prefix.js';
@@ -13,6 +14,7 @@ export default [
   fieldNameMappingSnakeCase,
   fieldOrder,
   forbidField,
+  forbidRequiredIgnoredField,
   modelNameGrammaticalNumber,
   modelNameMappingSnakeCase,
   modelNamePrefix,